demo_chart_timeframe_dropdown.py
#!/usr/bin/env python3
"""
Demonstration script for the enhanced chart/timeframe dropdown functionality.

This script shows how to use the new chart/timeframe dropdown detection
and control capabilities that have been added to handle the button
shown in your screenshot.
"""

from __future__ import annotations

import sys
import os

# Add utils to Python path
sys.path.append('utils')

try:
    from selenium_ui_controls import HighPriorityControls
    from capabilities.TF_dropdown_retract import TF_Dropdown_Retract
    from capabilities.base import Ctx
    print("✅ Successfully imported enhanced modules")
except Exception as e:
    print(f"❌ Import error: {e}")
    sys.exit(1)


def demo_high_priority_controls():
    """Demonstrate the new HighPriorityControls functionality."""
    print("\n🎯 HighPriorityControls Chart/Timeframe Dropdown Demo")
    print("-" * 55)

    # Note: This would normally use a real WebDriver
    # For demo purposes, we'll show the API structure
    print("""
📝 API Usage Example:

    from selenium_ui_controls import HighPriorityControls

    # Initialize with your WebDriver
    hpc = HighPriorityControls(driver)

    # 1. Find the chart/timeframe dropdown button
    find_meta = hpc.find_chart_timeframe_dropdown_with_meta()
    if find_meta['ok']:
        print(f"✅ Found button: {find_meta['selector_used']}")
        print(f"📊 Button details: {find_meta.get('button_text', 'N/A')}")

    # 2. Click the chart/timeframe dropdown button
    click_meta = hpc.click_chart_timeframe_dropdown_with_meta()
    if click_meta['ok']:
        print("✅ Successfully clicked dropdown button"        print(f"🔄 Click method: {click_meta.get('click_method', 'N/A')}")
        print(f"📂 Dropdown opened: {click_meta.get('dropdown_opened', False)}")

    # 3. Use convenience methods
    button = hpc.find_chart_timeframe_dropdown()
    success = hpc.click_chart_timeframe_dropdown()
    """)


def demo_tf_dropdown_retract():
    """Demonstrate the improved TF_Dropdown_Retract functionality."""
    print("\n🔄 TF_Dropdown_Retract Capability Demo")
    print("-" * 40)

    print("""
📝 API Usage Example:

    from capabilities.TF_dropdown_retract import TF_Dropdown_Retract
    from capabilities.base import Ctx

    # Initialize capability
    tf_retract = TF_Dropdown_Retract()
    ctx = Ctx(driver, verbose=True)

    # 1. Open the chart/timeframe dropdown
    open_result = tf_retract.run(ctx, {"action": "open"})
    if open_result.ok:
        print("✅ Dropdown opened successfully")

    # 2. Toggle dropdown (open then close)
    toggle_result = tf_retract.run(ctx, {"action": "toggle"})
    if toggle_result.ok:
        print("✅ Dropdown toggled successfully")

    # 3. Close dropdown
    close_result = tf_retract.run(ctx, {"action": "close"})
    if close_result.ok:
        print("✅ Dropdown closed successfully")
    """)


def demo_integration():
    """Show how the new functionality integrates with existing capabilities."""
    print("\n🔗 Integration with Existing Capabilities")
    print("-" * 45)

    print("""
📝 Enhanced topdown_select.py Usage:

    from capabilities.topdown_select import TopdownSelect

    # The capability now uses enhanced detection automatically
    topdown = TopdownSelect()
    result = topdown.run(ctx, {
        "stack": "1m",
        "labels": ["H1", "M15", "M5", "M1"],
        "save": True,
        "reopen_each": True
    })

    # Benefits:
    # ✅ Better button detection for your specific UI
    # ✅ More reliable dropdown opening/closing
    # ✅ Enhanced error reporting and debugging info
    # ✅ Backward compatible with existing code
    """)


def show_key_improvements():
    """Highlight the key improvements made."""
    print("\n🚀 Key Improvements Made")
    print("-" * 25)

    improvements = [
        "🎯 Added specific chart/timeframe dropdown button detection",
        "🔄 Fixed state persistence issues in TF_dropdown_retract.py",
        "📊 Enhanced selector strategies for better reliability",
        "🛠️ Improved error handling and metadata reporting",
        "🔧 Added convenience methods for easier usage",
        "⚡ Better integration with existing capabilities",
        "🧪 Comprehensive test coverage and validation"
    ]

    for improvement in improvements:
        print(f"  {improvement}")

    print("\n📋 Files Enhanced:")
    print("  • utils/selenium_ui_controls.py - New button detection methods")
    print("  • capabilities/TF_dropdown_retract.py - Fixed state management")
    print("  • capabilities/topdown_select.py - Enhanced integration")
    print("  • test_chart_timeframe_dropdown.py - Comprehensive tests")


def main():
    """Main demonstration function."""
    print("🎉 Enhanced Chart/Timeframe Dropdown Functionality Demo")
    print("=" * 60)

    # Show the improvements made
    show_key_improvements()

    # Demonstrate the APIs
    demo_high_priority_controls()
    demo_tf_dropdown_retract()
    demo_integration()

    # Usage instructions
    print("\n📚 Next Steps:")
    print("  1. Use the enhanced HighPriorityControls methods in your automation")
    print("  2. Replace manual dropdown handling with TF_Dropdown_Retract capability")
    print("  3. Update existing topdown_select usage to benefit from improvements")
    print("  4. Test with your specific trading platform UI")

    print("\n✨ The button from your screenshot should now be reliably detected!")
    print("   The dropdown will properly open and close as expected.")


if __name__ == "__main__":
    main()